# OAS_td_final_part1.yaml
openapi: 3.1.0
info:
  title: Agricultural federation API
  description: |
    # Versioning
    This API uses semantic versioning (MAJOR.MINOR.PATCH). Available versions are : 
    - v0.0.1 (2026-04-21) : POST /collectivities for collectivities creation and POST /members for members creation.
    - v0.0.2 (2026-04-22) : PUT /collectivities/{id}/informations for collectivity information update.
    - v0.0.3 (2026-04-23) : New features about membership fees, collectivities transactions and its financial accounts, and members payments.
    - v0.0.4 (2026-04-23) : GET /collectivities/{id} and  GET /collectivities/{id}/financialAccounts
    - v0.0.5 (2026-05-05) : Partial features collectivity statistics.
  version: 0.0.5
servers:
  - url: 'http://localhost:8080'
paths:
...
  '/collectivities/{id}/statistics':
    get:
      summary: Get local statistics for a collectivity
      description: |
        Get statistics for a collectivity between a date interval.
        Actually, statistics contain only earned amount from each member and potential unpaid amount.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
        - in: query
          name: from
          schema:
            type: string
            format: date
          required: true
        - in: query
          name: to
          schema:
            type: string
            format: date
          required: true
      responses:
        200:
          description: Collectivity local statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectivityLocalStatistics'
  '/collectivities/statistics':
    get:
      summary: Get collectivities overall statistics
      description: |
        Get collectivities global statistics between a date interval.
        Actually, statistics contain only overall percentage of members who are current with their dues (membership fees) and number of new members.
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
          required: true
        - in: query
          name: to
          schema:
            type: string
            format: date
          required: true
      responses:
        200:
          description: Collectivity overall statistics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CollectivityOverallStatistics'
components:
  schemas:
...
    CollectivityInformation:
      type: object
      properties:
        name:
          type: string
        number:
          type: int
...
    CollectivityLocalStatistics:
      type: object
      properties:
        memberDescription:
          $ref: '#/components/schemas/MemberDescription'
        earnedAmount:
          type: number
        unpaidAmount:
          type: number
    CollectivityOverallStatistics:
      type: object
      properties:
        collectivityInformation:
          $ref: '#/components/schemas/CollectivityInformation'
        newMembersNumber:
          type: integer
        overallMemberCurrentDuePercentage:
          type: number
...
    MemberDescription:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        occupation:
          type: string
...